backend/requirements.txt
flask
flask-cors
pycryptodome
